requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
psycopg[binary]
gunicorn==23.0.0
Werkzeug==3.1.4
whitenoise-6.11.0
